AGENT/BookTemplateEngine/requirements.txt
openai
python-dotenv
python-docx
PyPDF2
flask
